requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.7
Django==5.0.1
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
exceptiongroup==1.1.2
fabric==3.2.2
filelock==3.12.2
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.3
MarkupSafe==2.1.3
mistune==3.0.2
multidict==6.0.4
mysqlclient==2.1.1
nose==1.3.7
nose2==0.12.0
numpy==1.23.4
oauthlib==3.2.2
packaging==23.1
paramiko==3.3.1
pep8==1.7.1
phonenumbers==8.13.36
pillow==10.3.0
pipenv==2023.11.15
platformdirs==3.10.0
pluggy==1.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pycparser==2.21
PyNaCl==1.5.0
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
tomli==2.0.1
tweepy==4.14.0
typing_extensions==4.5.0
tzdata==2023.4
urllib3==2.0.7
virtualenv==20.24.2
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.4
